[refactor] 감상 서버 통신 관련 기초 세팅 / 패키지 및 네이밍 변경 (#102)
